# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger Docs",
    "description": "Docs for app backend",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3003",
  "basePath": "/api",
  "tags": [
    {
      "name": "Restaurants",
      "description": "API for restaurants in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/restaurant": {
      "post": {
        "tags": [
          "Restaurants"
        ],
        "description": "Create new restaurant in system",
        "parameters": [
          {
            "name": "restaurant",
            "in": "body",
            "description": "Restaurant that we want to create",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New restaurant is created",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Restaurants"
        ],
        "summary": "Get all restaurant in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Restaurants"
            }
          }
        }
      }
    },
    "/restaurant/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Restaurants"
        ],
        "summary": "Get restaurant with given ID",
        "responses": {
          "200": {
            "description": "Restaurant is found",
            "schema": {
              "$ref": "#/definitions/Restaurant"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user with given ID",
        "tags": [
          "Restaurants"
        ],
        "responses": {
          "200": {
            "description": "Restaurant is deleted"
          }
        }
      }
    },
    "/menu": {
      "post": {
        "tags": [
          "Menus"
        ],
        "description": "Create new menu in system",
        "parameters": [
          {
            "name": "menu",
            "in": "body",
            "description": "Menu that we want to create",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New menu is created",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Menus"
        ],
        "summary": "Get all menus in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Menus"
            }
          }
        }
      }
    },
    "/menu/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of menu that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Menus"
        ],
        "summary": "Get menu with given ID",
        "responses": {
          "200": {
            "description": "Menu is found",
            "schema": {
              "$ref": "#/definitions/Menu"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete menu with given ID",
        "tags": [
          "Menus"
        ],
        "responses": {
          "200": {
            "description": "Menu is deleted"
          }
        }
      }
    },
    "/item": {
      "post": {
        "tags": [
          "Items"
        ],
        "description": "Create new item in system",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "description": "Item that we want to create",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New item is created",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get all items in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Items"
            }
          }
        }
      }
    },
    "/item/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of item that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get item with given ID",
        "responses": {
          "200": {
            "description": "Item is found",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete item with given ID",
        "tags": [
          "Items"
        ],
        "responses": {
          "200": {
            "description": "Item is deleted"
          }
        }
      }
    }
  },
  "definitions": {
    "Restaurant": {
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        },

        "cuisine": {
          "type": "string"
        },

        "isVegetarian": {
          "type": "boolean"
        },
        "image": {
          "type": "string"
        },
        "rating": {
          "type": "number"
        },
        "location": {
          "type": "object"
        }
      }
    },
    "Restaurants": {
      "type": "array",
      "$ref": "#/definitions/Restaurant"
    },
    "Item": {
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "cuisine": {
          "type": "string"
        },
        "isVegetarian": {
          "type": "boolean"
        },
        "image": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "label": {
          "type": "string"
        },
        "rating": {
          "type": "number"
        }
      }
    },
    "Items": {
      "type": "array",
      "$ref": "#/definitions/Item"
    },
    "Menu": {
      "properties": {
        "restaurantId": {
          "type": "string"
        },
        "itemIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Menus": {
      "type": "array",
      "$ref": "#/definitions/Menu"
    }
  }
}